# swagger.yaml
openapi: 3.0.3
info:
  title: Equipment Loan Backend API
  version: 1.0.0
  description: Expressで実装された備品貸出API
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: メールアドレスとパスワードでログイン
      description: 同一アカウントで再ログインすると、それ以前に発行されたトークンは無効化される
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: 入力不備
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                validationError:
                  value:
                    message: email and password are required
                invalidEmail:
                  value:
                    message: email must be a valid address
                shortPassword:
                  value:
                    message: password must be at least 8 characters
        "401":
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidCredentials:
                  value:
                    message: invalid credentials
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: ログアウト
      description: bearerAuth を使ってログアウトする。実行後はそれまでのトークンが無効化される
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ログアウト成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
              examples:
                success:
                  value:
                    ok: true
                    message: logged out
        "401":
          description: 未認証または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidated:
                  value:
                    message: token expired by newer login or logout
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required:
        - employeeId
        - email
        - name
        - department
        - role
      properties:
        employeeId:
          type: string
          example: EMP-0142
        email:
          type: string
          format: email
          example: misaki.tanaka@example.com
        name:
          type: string
          example: 田中 美咲
        department:
          type: string
          example: プロダクト開発部
        role:
          type: string
          example: member
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: misaki.tanaka@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: Passw0rd!
    LoginResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: string
          example: 1h
        user:
          $ref: "#/components/schemas/User"
    LogoutResponse:
      type: object
      required:
        - ok
        - message
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: logged out
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: equipment not found